requirements.txt
customtkinter
numpy
opencv_python
Pillow
pycryptodome
PyWavelets
